Add files via upload
This is where we get hands on with reinforcement learning and human feedback, or RLHF. The purpose of this lab is to lower the toxicity of our instruction fine-tuned model from Lab 2, that was the output of Lab 2. That'll be the input to Lab 3 here. And we will lower the toxicity using a hate speech reward model where we want to optimize for not hate. We'll be using PPO that you learned about in the lesson, and we will wrap everything up with quantitative and then qualitative comparison of our detoxification process. Okay, let's install our Python libraries. Here we see that we're using PyTorch. We're also using the same transformers that we used in the previous labs, of course, data sets to get access to our public data set, evaluate so that we could run the rouge_score. Peft is the parameter efficient fine-tuning library. And here's a new library called trl. And this is what's going to give us access to PPO. And a similar type of trainer and then training arguments, there's going to be PPOTrainer and then training arguments following the hugging face convention of trainer and training arguments. Let's do a bunch of our imports up here. Familiar classes, the AutoModelForSeq2Seq. Note that there's a new one here, which is AutoModelForSeq1Classification. This is what we're going to use to load our Facebook binary classifier, seq1 classifier. Which when we give it a string of text or a sequence of text, it'll tell us whether or not that text contains hate speech or not, with a particular distribution across not hate or hate. And of course, in this lab, we'll be optimizing for not hate. Here we see our Peft and our LoraConfig, which we saw in Lab 2. Here we see the specific PPOTrainer. There's also this other class, AutoModelForSeqtoSeq language model with a value head. This is what's required when we do the PPO training, which we'll see in a bit. Also, this LengthSampler. This will give us the ability to filter or actually sample from our text various lengths, so we don't actually have to pull in the entire sequence, we could pull in just different samples.